(feat) Consistent ActivityStream authentication with directory-sso
This _is_ mostly just a copy+paste job, bringing in

  - Mitigation against X-Forwarded-For spoofing
  - Signing the requests

It may be reaching the point where it's valuable to move this out to a separate
library, and we have confidence that it doesn't need to handle lots of
different cases where it would have been better to just keep them separate. But...

 - Different projects are using different Python versions

 - I think there is still value in having the tests repeated in each:
   combinations of middleware can have a bit impact on what's happening in
   views with regards to authentication. The tests are slightly different due to
   routing differences.

 - Different projects have different cache/redis configs (some without replay
   attack prevention)

 - There could be differences introduced when we actually start returning data

Since I don't see evidence it would be harder to do later, so prefer to do
later.

Also, different projects have slightly different configuration of caching (i.e.
some have Redis and some not).